fix tensorrt warning 'Your ONNX model has been generated with INT64 weights, while TensorRT does not natively support INT64. Attempting to cast down to INT32.' by changing model int64 parameter to int32

+++ rl/models.py
@@ -73,14 +73,14 @@ class EnvEmbedding(nn.Module):
                           (MAX_PLAYER_NUMBER - 1, num_bins),
                           (MAX_PLAYER_NUMBER - 1, num_bins + 1),
                           ]
-        self.starter_idx_array = nn.Parameter(torch.arange(0, num_starters, dtype=torch.int16), requires_grad=False)
+        self.starter_idx_array = nn.Parameter(torch.arange(0, num_starters, dtype=torch.int32), requires_grad=False)
         current_start_idx = num_starters
         field_start_idx_list = list()
         for num_fields, num_dims in field_dim_list:
             for _ in range(num_fields):
                 field_start_idx_list.append(current_start_idx)
             current_start_idx += num_dims
-        self.field_start_idx_array = nn.Parameter(torch.tensor(field_start_idx_list, dtype=torch.int16), requires_grad=False)
+        self.field_start_idx_array = nn.Parameter(torch.tensor(field_start_idx_list, dtype=torch.int32), requires_grad=False)
         self.field_embedding = nn.Embedding(num_embeddings=num_starters + sum(item[1] for item in field_dim_list), embedding_dim=embedding_dim)
 
         # card_segments = [0, 0, 1, 1, 1, 2, 3, 4, 4, 4, 4, 4, 4, 4]
@@ -114,7 +114,7 @@ class EnvEmbedding(nn.Module):
             # self.card_embedding_idx = nn.Parameter(torch.cat([card_embedding_idx_1, card_embedding_idx_2, card_embedding_idx_3], dim=0).unsqueeze(0), requires_grad=False)
 
             self.position_embedding = nn.Embedding(num_embeddings=num_starters + self.embedding_sequence_len, embedding_dim=self.positional_embedding_dim)
-            self.position_embedding_idx = nn.Parameter(torch.arange(num_starters + self.embedding_sequence_len, dtype=torch.int16).unsqueeze(0), requires_grad=False)
+            self.position_embedding_idx = nn.Parameter(torch.arange(num_starters + self.embedding_sequence_len, dtype=torch.int32).unsqueeze(0), requires_grad=False)
 
     def forward(self, x):
         batch_size = x.shape[0]
